+++ Engine.Core/Managers/EventManager.cs
@@ -19,7 +19,6 @@ namespace Engine.Core.Managers
         private readonly DiagnosticsController _diagnosticsController;
 
         public ConcurrentQueue<EngineEvt> EventQueue { get; private set; }
-        private bool Running { get; set; } = true;
 
         public ConcurrentDictionary<EventType, ConcurrentBag<IEventSubscriber>> Subscribers { get; set; }
 
@@ -32,9 +31,9 @@ namespace Engine.Core.Managers
             Subscribers = new ConcurrentDictionary<EventType, ConcurrentBag<IEventSubscriber>>();
         }
 
-        public async void Begin()
+        public async void Run()
         {
-            while (Running)
+            while (EngineStatics.Running)
             {
                 await ProcessEvents();
             }
@@ -54,33 +53,11 @@ namespace Engine.Core.Managers
             if (EventQueue.TryDequeue(out EngineEvt engineEvt))
             {
                 _logger.LogInformation("Found event '{EventType}'. Adding publish task for subscribers.", engineEvt.EventType);
-                Subscribers[engineEvt.EventType].ToList().ForEach(x => tasks.Add(x.HandleEvent(engineEvt)));
+                if (Subscribers.ContainsKey(engineEvt.EventType))
+                    Subscribers[engineEvt.EventType].ToList().ForEach(x => tasks.Add(x.HandleEvent(engineEvt)));
             }
 
             await _diagnosticsController.DiagnoseTask(Task.WhenAll(tasks), "HandleEvent");
-        }
-
-        private object GetTypedEvent(EngineEvt engineEvt)
-        {
-            switch (engineEvt.EventType)
-            {
-                case EventType.KEYBOARD:
-                    return (KeyboardEvt)engineEvt;
-                default:
-                    _logger.LogError("Event type {EventType} not supported", engineEvt.EventType);
-                    break;
-            }
-            return null;
-        }
-
-        public bool OnOffSwitch(bool? @switch)
-        {
-            if (@switch.HasValue)
-                Running = @switch.Value;
-            else
-                Running = !Running;
-
-            return Running;
         }
 
         public Task SubscribeToEvent(EventType eventType, IEventSubscriber subscriber)