+++ Engine.Core/Managers/Interfaces/IEventManager.cs
@@ -9,7 +9,7 @@ namespace Engine.Core.Managers.Interfaces
 {
     public interface IEventManager
     {
-        void Begin();
+        void Run();
         Task PublishEvent<T>(T @event) where T : EngineEvt;
         Task SubscribeToEvent(EventType eventType, IEventSubscriber subscriber);
         Task UnsubscribeFromEvent(EventType eventType, IEventSubscriber subscriber);